+++ Source_Files/RenderOther/Shape_Blitter.cpp
@@ -214,10 +214,10 @@ void Shape_Blitter::OGL_Draw(const Image_Rect& dst)
             U_Scale *= crop_rect.h / static_cast<double>(m_scaled_src.h);
 
 		GLfloat texcoords[8] = {
-			U_Offset, V_Offset,
-			U_Offset, V_Offset + V_Scale,
-			U_Offset + U_Scale, V_Offset + V_Scale,
-			U_Offset + U_Scale, V_Offset
+			GLfloat(U_Offset),           GLfloat(V_Offset),
+			GLfloat(U_Offset),           GLfloat(V_Offset + V_Scale),
+			GLfloat(U_Offset + U_Scale), GLfloat(V_Offset + V_Scale),
+			GLfloat(U_Offset + U_Scale), GLfloat(V_Offset)
 		};
 		GLfloat vertices[8] = {
 			dst.x, dst.y,
@@ -337,8 +337,8 @@ void Shape_Blitter::SDL_Draw(SDL_Surface *dst_surface, const Image_Rect& dst)
     if (!m_scaled_surface)
         return;
     
-    SDL_Rect r = { crop_rect.x, crop_rect.y, crop_rect.w, crop_rect.h };
-    SDL_Rect sdst = { dst.x, dst.y, dst.w, dst.h };
+	SDL_Rect r = { int(crop_rect.x), int(crop_rect.y), int(crop_rect.w), int(crop_rect.h) };
+	SDL_Rect sdst = { int(dst.x), int(dst.y), int(dst.w), int(dst.h) };
 	SDL_BlitSurface(m_scaled_surface, &r, dst_surface, &sdst);
 }
